+++ src/main/java/acme/features/authenticated/canParticipate/AuthenticatedCanParticipateController.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import acme.components.CustomCommand;
 import acme.entities.messageThreads.CanParticipate;
 import acme.framework.components.BasicCommand;
 import acme.framework.controllers.AbstractController;
@@ -17,16 +18,20 @@ import acme.framework.entities.Authenticated;
 public class AuthenticatedCanParticipateController extends AbstractController<Authenticated, CanParticipate> {
 
 	@Autowired
-	private AuthenticatedCanParticipateCreateService	createService;
+	private AuthenticatedCanParticipateListInvolvedService	listInvolvedService;
 
 	@Autowired
-	private AuthenticatedCanParticipateDeleteService	deleteService;
+	private AuthenticatedCanParticipateShowService			showService;
+
+	@Autowired
+	private AuthenticatedCanParticipateDeleteService		deleteService;
 
 
 	@PostConstruct
 	private void initialise() {
-		super.addBasicCommand(BasicCommand.CREATE, this.createService);
+		super.addCustomCommand(CustomCommand.LIST_INVOLVED, BasicCommand.LIST, this.listInvolvedService);
 		super.addBasicCommand(BasicCommand.DELETE, this.deleteService);
+		super.addBasicCommand(BasicCommand.SHOW, this.showService);
 	}
 
 }
